refactor(operator_commands_dashboard): extract current_work + cycle helpers
Move current_work HTTP handler plus format_recent_actions_for_cycle and
read_recent_cycle_reports helpers from routes.rs into a new sibling
current_work.rs module. Update activity.rs and workboard.rs to import
the moved helpers from current_work instead of routes.

- routes.rs: 4395 -> 4110
- current_work.rs (NEW, 295)

Refs #1266
