+++ src/operator_commands_dashboard/workboard.rs
@@ -1,11 +1,12 @@
 use axum::Json;
 use serde_json::{Value, json};
 
-use super::routes::{is_pid_alive, read_recent_cycle_reports, resolve_state_root};
+use super::routes::{is_pid_alive, resolve_state_root};
+use super::current_work::read_recent_cycle_reports;
 use crate::agent_registry::{AgentRegistry, FileBackedAgentRegistry};
 use crate::goal_curation::GoalBoard;
 use crate::cognitive_memory::{CognitiveMemoryOps, NativeCognitiveMemory};
-use super::routes::format_recent_actions_for_cycle;
+use super::current_work::format_recent_actions_for_cycle;
 
 // ---------------------------------------------------------------------------
 // Workboard API — aggregated view of Simard's current mental state
